# Optimize timestamp run-length decoding