--- a/tsdb/engine/tsm1/timestamp.go
+++ b/tsdb/engine/tsm1/timestamp.go
@@ -184,11 +184,16 @@ func (e *encoder) encodeRLE(first, delta, div uint64, n int) ([]byte, error) {
 }
 
 type TimeDecoder struct {
-	v   int64
-	i   int
-	ts  []uint64
-	dec simple8b.Decoder
-	err error
+	v    int64
+	i, n int
+	ts   []uint64
+	dec  simple8b.Decoder
+	err  error
+
+	// The delta value for a run-length encoded byte slice
+	rleDelta int64
+
+	encoding byte
 }
 
 func (d *TimeDecoder) Init(b []byte) {
@@ -196,10 +201,23 @@ func (d *TimeDecoder) Init(b []byte) {
 	d.i = 0
 	d.ts = d.ts[:0]
 	d.err = nil
+	if len(b) > 0 {
+		// Encoding type is stored in the 4 high bits of the first byte
+		d.encoding = b[0] >> 4
+	}
 	d.decode(b)
 }
 
 func (d *TimeDecoder) Next() bool {
+	if d.encoding == timeCompressedRLE {
+		if d.i >= d.n {
+			return false
+		}
+		d.i++
+		d.v += d.rleDelta
+		return d.i < d.n
+	}
+
 	if d.i >= len(d.ts) {
 		return false
 	}
@@ -221,9 +239,7 @@ func (d *TimeDecoder) decode(b []byte) {
 		return
 	}
 
-	// Encoding type is stored in the 4 high bits of the first byte
-	encoding := b[0] >> 4
-	switch encoding {
+	switch d.encoding {
 	case timeUncompressed:
 		d.decodeRaw(b[1:])
 	case timeCompressedRLE:
@@ -231,7 +247,7 @@ func (d *TimeDecoder) decode(b []byte) {
 	case timeCompressedPackedSimple:
 		d.decodePacked(b)
 	default:
-		d.err = fmt.Errorf("unknown encoding: %v", encoding)
+		d.err = fmt.Errorf("unknown encoding: %v", d.encoding)
 	}
 }
 
@@ -279,20 +295,11 @@ func (d *TimeDecoder) decodeRLE(b []byte) {
 	// Last 1-10 bytes is how many times the value repeats
 	count, _ := binary.Uvarint(b[i:])
 
-	// Rebuild construct the original values now
-	deltas := d.ts[:0]
-	for i := 0; i < int(count); i++ {
-		deltas = append(deltas, value)
-	}
+	d.v = int64(first - value)
+	d.rleDelta = int64(value)
 
-	// Reverse the delta-encoding
-	deltas[0] = first
-	for i := 1; i < len(deltas); i++ {
-		deltas[i] = deltas[i-1] + deltas[i]
-	}
-
-	d.i = 0
-	d.ts = deltas
+	d.i = -1
+	d.n = int(count)
 }
 
 func (d *TimeDecoder) decodeRaw(b []byte) {