@jwilder 2016-05-20
Still used by simple8b decoder. I'll remove that next.